Remove types package
