+++ packages/common/package.json
@@ -12,7 +12,6 @@
   "typings": "dist/index.d.ts",
   "devDependencies": {
     "@bn-onboard/common": "^0.0.1",
-    "@bn-onboard/types": "^0.0.1",
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^11.0.0",
@@ -24,6 +23,7 @@
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-svelte3": "^3.2.1",
+    "eventemitter3": "^4.0.7",
     "prettier": "^2.4.0",
     "prettier-plugin-svelte": "^2.4.0",
     "rollup": "^2.3.4",
@@ -35,6 +35,7 @@
     "typescript": "4.3.5"
   },
   "dependencies": {
+    "@ethersproject/providers": "^5.5.0",
     "joi": "^17.4.2",
     "rxjs": "^7.3.0"
   }
