+++ packages/gnosis/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@bn-onboard/common": "^0.0.1",
-    "@bn-onboard/types": "^0.0.1",
     "@gnosis.pm/safe-apps-provider": "^0.9.2",
     "@gnosis.pm/safe-apps-sdk": "^6.1.1"
   }
