+++ packages/injected/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@bn-onboard/common": "^0.0.1",
-    "@bn-onboard/types": "^0.0.1",
     "joi": "^17.4.2",
     "lodash.uniqby": "^4.7.0"
   }
