+++ packages/mew/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@bn-onboard/common": "^0.0.1",
-    "@bn-onboard/types": "^0.0.1",
     "@myetherwallet/mewconnect-web-client": "^2.2.0-beta.14",
     "rxjs": "^7.4.0"
   }
