+++ packages/torus/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@bn-onboard/common": "^0.0.1",
-    "@bn-onboard/types": "^0.0.1",
     "@toruslabs/torus-embed": "^1.18.3"
   }
 }
