--- packages/types/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "@bn-onboard/types",
-  "version": "0.0.1",
-  "description": "Shared Typescript types",
-  "module": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "yarn.lock"
-  ],
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc -w",
-    "test": "ava"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "@ethersproject/providers": "^5.5.0"
-  },
-  "devDependencies": {
-    "eventemitter3": "^4.0.7",
-    "typescript": "^4.4.4"
-  }
-}
