--- packages/types/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*"],
-  
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src",
-    "declarationDir": "dist",
-    "paths": {
-      "*": ["./src/*", "./node_modules/*"]
-    },
-    "typeRoots": ["node_modules/@types"]
-  }
-}
