+++ packages/walletconnect/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@bn-onboard/common": "^0.0.1",
-    "@bn-onboard/types": "^0.0.1",
     "@walletconnect/client": "^1.6.6",
     "@walletconnect/qrcode-modal": "^1.6.6",
     "rxjs": "^7.4.0"
